requirements.txt
streamlit
torch
torchvision
diffusers
transformers
opencv-python
numpy
Pillow
